+++ packages/ecs-helpers/lib/index.js
@@ -21,7 +21,7 @@ const errorFormatters = require('./error-formatters')
 const httpFormatters = require('./http-formatters')
 
 module.exports = {
-  version: '1.6.0',
+  version: '8.10.0',
   ...errorFormatters,
   ...httpFormatters
 }