+++ packages/ecs-helpers/test/hapi.test.js
@@ -65,8 +65,10 @@ test('hapi res/req serialization', testOpts, t => {
         headers: {
           'user-agent': 'cool-agent',
           host: `127.0.0.1:${server.info.port}`,
-          connection: 'close'
-        }
+          connection: 'close',
+          'request-id': 'arequestid'
+        },
+        id: 'arequestid'
       },
       response: {
         status_code: 200,
@@ -98,8 +100,16 @@ test('hapi res/req serialization', testOpts, t => {
     t.comment('hapi server running on %s', server.info.uri)
 
     // Make a request so we trigger a 'response' event above.
-    const req = http.get(`http://127.0.0.1:${server.info.port}/`,
-      { headers: { 'user-agent': 'cool-agent', connection: 'close' } })
+    const req = http.get(
+      `http://127.0.0.1:${server.info.port}/`,
+      {
+        headers: {
+          'user-agent': 'cool-agent',
+          connection: 'close',
+          'request-id': 'arequestid'
+        }
+      }
+    )
     req.on('error', t.ifErr)
   })
 })